debian/ubuntu scripts - make sure we're running /usr/bin/perl + rebuild

+++ scripts/build_debian.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 cd .. && grep -s -q APP-MTAWS-ROOT-DIR Build.PL || exit 1 # protect from wrong dir
-
+[ `which perl` = "/usr/bin/perl" ] || exit 1
+perl Build.PL && ./Build build
 rm -rf debian
 mkdir -p debian/source
 scripts/write_debian_files.pl ./packaging debian wheezy
